+++ EtradeSystem/src/Domain/Users/ExternalService/Payment/IPay.java
@@ -3,6 +3,6 @@ package Domain.Users.ExternalService.Payment;
 import java.time.LocalTime;
 
 public interface IPay {
-    public boolean pay(int cardFrom, LocalTime expDate, int cvv,int price, int cardTo);
-    public boolean canPay(int cardFrom, LocalTime expDate, int cvv,int price);
+    public boolean pay(int cardFrom, LocalTime expDate, int cvv,double price, int cardTo);
+    public boolean canPay(int cardFrom, LocalTime expDate, int cvv,double price);
 }
